+++ one-ui/src/main/java/com/marklogic/hub/oneui/models/HubConfigSession.java
@@ -881,7 +881,7 @@ public class HubConfigSession implements HubConfig, InitializingBean, Disposable
      */
     @Override
     public DatabaseClient newReverseFlowClient() {
-        return getClientByDatabaseKindAndName(DatabaseKind.STAGING, hubConfigImpl.getDbName(DatabaseKind.FINAL));
+        return getClientByDatabaseKindAndName(DatabaseKind.FINAL, hubConfigImpl.getDbName(DatabaseKind.STAGING));
     }
 
     /**
@@ -1040,5 +1040,10 @@ public class HubConfigSession implements HubConfig, InitializingBean, Disposable
     //only for test purpose
     protected Map<DatabaseKind, Map<String, DatabaseClient>> getAllDatabaseClients() {
         return clientsByKindAndDatabaseName;
+    }
+
+    // this is used for flow runner, so we have a HubConfig that isn't session scoped
+    public HubConfigImpl getHubConfigImpl() {
+        return hubConfigImpl;
     }
 }