+++ one-ui/src/test/java/com/marklogic/hub/curation/controllers/FlowControllerTest.java
@@ -7,11 +7,15 @@ import com.marklogic.client.document.GenericDocumentManager;
 import com.marklogic.client.io.StringHandle;
 import com.marklogic.hub.ApplicationConfig;
 import com.marklogic.hub.flow.Flow;
+import com.marklogic.hub.flow.RunFlowResponse;
 import com.marklogic.hub.flow.impl.FlowImpl;
+import com.marklogic.hub.flow.impl.FlowRunnerImpl;
 import com.marklogic.hub.oneui.Application;
 import com.marklogic.hub.oneui.TestHelper;
 import com.marklogic.hub.oneui.models.HubConfigSession;
 import com.marklogic.hub.oneui.models.StepModel;
+import org.apache.commons.io.FileUtils;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -21,9 +25,13 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.io.IOException;
+import java.nio.file.Paths;
+import java.util.Collections;
 import java.util.List;
 
 
@@ -41,13 +49,13 @@ class FlowControllerTest {
             "  \"steps\" : { }\n" +
             "}";
 
-    private final String stepString = "{\"name\":\"e2e-xml\",\"description\":\"\",\"options\":{\"additionalCollections\":[]" +
+    private final String stepString = "{\"name\":\"e2e-json\",\"description\":\"\",\"options\":{\"additionalCollections\":[]" +
             ",\"headers\":{\"sources\":[{\"name\":\"runXqyFuncFlow\"}],\"createdOn\":\"currentDateTime\",\"createdBy\":\"currentUser\"}," +
             "\"sourceQuery\":\"\",\"collections\":[\"xml-xqy\"],\"permissions\":\"data-hub-operator,read,data-hub-operator,update\"," +
             "\"outputFormat\":\"xml\",\"targetDatabase\":\"data-hub-STAGING\"},\"customHook\":{},\"retryLimit\":0,\"batchSize\":0," +
             "\"threadCount\":0,\"stepDefinitionName\":\"default-ingestion\",\"stepDefinitionType\":\"INGESTION\",\"fileLocations\":" +
-            "{\"inputFilePath\":\"/Users/ssambasu/source/marklogic-data-hub/marklogic-data-hub/ye-olde-project/input\",\"inputFileType\":" +
-            "\"xml\",\"outputURIReplacement\":\".*/input,'/xqyfunc'\",\"separator\":\"\"}}";
+            "{\"inputFilePath\":\"input\",\"inputFileType\":" +
+            "\"json\",\"outputURIReplacement\":\".*/input,'/xqyfunc'\",\"separator\":\"\"}}";
 
     private final String customStepString = "{\"name\":\"second\",\"description\":\"\",\"isValid\":false,\"modulePath\":\"\",\"options\":" +
             "{\"collections\":[\"second\"],\"additionalCollections\":[],\"sourceQuery\":\"cts.collectionQuery([])\",\"sourceCollection\":\"\"," +
@@ -56,13 +64,16 @@ class FlowControllerTest {
             "\"user\":\"\",\"runBefore\":false},\"batchSize\":100,\"threadCount\":4,\"stepDefinitionType\":\"CUSTOM\",\"stepDefType\":\"CUSTOM\"," +
             "\"stepDefinitionName\":\"second\",\"selectedSource\":\"\"}";
 
-    JsonNode validLoadDataConfig = new ObjectMapper().readTree("{ \"name\": \"validArtifact\", \"sourceFormat\": \"xml\", \"targetFormat\": \"json\"}");
+    JsonNode validLoadDataConfig = new ObjectMapper().readTree("{ \"name\": \"validArtifact\", \"sourceFormat\": \"json\", \"targetFormat\": \"json\", \"inputFilePath\": \"input\"}");
 
     @Autowired
-    private FlowController controller;
+    private FlowTestController controller;
 
     @Autowired
     private LoadDataController loadDataController;
+
+    @Autowired
+    private JobsController jobsController;
 
     @Autowired
     private TestHelper testHelper;
@@ -78,6 +89,20 @@ class FlowControllerTest {
     @BeforeEach
     void before(){
         testHelper.authenticateSession();
+        testHelper.setHubProjectDirectory();
+    }
+
+    @AfterEach
+    void after(){
+        ResponseEntity<?> flowResp = null;
+        try {
+            flowResp = controller.getFlow("testFlow");
+        } catch (Exception e) {
+
+        }
+        if (flowResp != null && flowResp.getBody() != null) {
+            controller.deleteFlow("testFlow");
+        }
     }
 
     @Test
@@ -119,7 +144,7 @@ class FlowControllerTest {
             Assertions.assertEquals(1, (steps.size()));
 
             //PUT step
-            controller.createStep("testFlow","e2e-xml-ingestion", mapper.writeValueAsString(stepModel));
+            controller.createStep("testFlow","e2e-json-ingestion", mapper.writeValueAsString(stepModel));
             //POST custom step
             controller.createStep("testFlow", 2, customStepString).getBody();
 
@@ -131,29 +156,29 @@ class FlowControllerTest {
 
             //link artifact to step options
             loadDataController.updateArtifact("validArtifact", validLoadDataConfig);
-            controller.linkArtifact("testFlow", "e2e-xml-ingestion", "loadData", "validArtifact");
+            controller.linkArtifact("testFlow", "e2e-json-ingestion", "loadData", "validArtifact");
 
             //GET step
-            stepModel = (StepModel)controller.getStep("testFlow", "e2e-xml-ingestion").getBody();
-            Assertions.assertEquals("e2e-xml", stepModel.getName());
+            stepModel = (StepModel)controller.getStep("testFlow", "e2e-json-ingestion");
+            Assertions.assertEquals("e2e-json", stepModel.getName());
             Assertions.assertEquals(100, stepModel.getBatchSize().intValue());
             // step should have LoadData link
             Assertions.assertTrue(stepModel.getOptions().has("loadData"), "Should have loadData link");
             Assertions.assertEquals("validArtifact", stepModel.getOptions().get("loadData").get("name").asText(), "Link should have expected name");
 
             // remove artifact link
-            controller.removeLinkToArtifact("testFlow", "e2e-xml-ingestion", "loadData", "validArtifact");
-            stepModel = (StepModel)controller.getStep("testFlow", "e2e-xml-ingestion").getBody();
+            controller.removeLinkToArtifact("testFlow", "e2e-json-ingestion", "loadData", "validArtifact");
+            stepModel = (StepModel)controller.getStep("testFlow", "e2e-json-ingestion");
             Assertions.assertFalse(stepModel.getOptions().has("loadData"), "Should not have loadData link");
 
             //DELETE step
-            controller.deleteStep("testFlow", "e2e-xml-ingestion" );
+            controller.deleteStep("testFlow", "e2e-json-ingestion" );
             controller.deleteStep("testFlow", "second-custom" );
 
             //the module should be deleted
             Assertions.assertFalse(docMgr.read("/step-definitions/custom/second/second.step.json").hasNext());
             try{
-                controller.getStep("testFlow", "e2e-xml-ingestion");
+                controller.getStep("testFlow", "e2e-json-ingestion");
                 Assertions.fail();
             }
             catch (Exception e){
@@ -163,13 +188,49 @@ class FlowControllerTest {
             //DELETE flow
             controller.deleteFlow("testFlow");
             try{
-                controller.getFlow("testFlow");
-                Assertions.fail();
+                Flow flow = (Flow) controller.getFlow("testFlow").getBody();
+                Assertions.assertNull(flow, "Flow shouldn't exist anymore");
             }
             catch (Exception e) {
                 logger.info("Exception is expected as the flow being fetched has been deleted");
             }
-            loadDataController.deleteLoadDataConfig("validArtifact");
+            loadDataController.deleteArtifact("validArtifact");
+        }
+    }
+
+    @Test
+    void runFlow() throws IOException {
+        ObjectMapper mapper = new ObjectMapper();
+        String hubProjectDir = hubConfig.getProjectDir();
+        FileUtils.copyDirectory(testHelper.getResourceFile("input"), Paths.get(hubProjectDir,"input").toFile());
+        controller.createFlow(flowString);
+        //PUT step
+        controller.createStep("testFlow", 1, stepString).getBody();
+        //link artifact to step options
+        loadDataController.updateArtifact("validArtifact", validLoadDataConfig);
+        controller.linkArtifact("testFlow", "e2e-json-ingestion", "loadData", "validArtifact");
+
+        RunFlowResponse resp = (RunFlowResponse) controller.runFlow("testFlow", Collections.singletonList("e2e-json-ingestion"));
+
+        Assertions.assertEquals("testFlow", resp.getFlowName(), "Run flow response has correct flow name");
+        controller.getLastFlowRunner().awaitCompletion();
+        JsonNode jobsDoc = jobsController.getJob(resp.getJobId());
+        String jobStatus = jobsDoc.get("jobStatus").asText();
+        Assertions.assertEquals("finished", jobStatus, "Job status should be 'finished' once threads complete");
+    }
+
+    @Controller
+    @RequestMapping("/api/test/flows")
+    static class FlowTestController extends FlowController {
+        protected FlowRunnerImpl lastFlowRunner = null;
+
+        protected FlowRunnerImpl getFlowRunner() {
+            lastFlowRunner = (FlowRunnerImpl) super.getFlowRunner();
+            return lastFlowRunner;
+        }
+
+        protected FlowRunnerImpl getLastFlowRunner() {
+            return lastFlowRunner;
         }
     }
 }