+++ one-ui/ui/src/App.tsx
@@ -11,7 +11,6 @@ import Home from './pages/Home';
 import Install from './pages/Install';
 import ProjectInfo from './pages/ProjectInfo';
 import LoadData from './pages/LoadData';
-import EntityTiles from './components/entities/entity-tiles';
 import Bench from './pages/Bench';
 import Reset from './pages/Reset';
 import NoMatchRedirect from './pages/noMatchRedirect';
@@ -23,6 +22,7 @@ import './App.scss';
 import Application from './config/application.config';
 import { themes, themeMap } from './config/themes.config';
 import axios from 'axios';
+import EntityTypes from './pages/EntityTypes';
 
 
 interface Props extends RouteComponentProps<any> {}
@@ -50,14 +50,14 @@ const App: React.FC<Props> = ({history, location}) => {
       }
       if (location.state && !user.redirect && user.error.type === '') {
         if (location.state.hasOwnProperty('from')) {
-          history.push(location.state.from.pathname);
+          history.push(location.state['from'].pathname);
         }
       }
       if (user.redirect || location.pathname === '/') {
         if (localStorage.getItem('dhIsInstalled') === 'false' && localStorage.getItem('dhUserHasManagePrivileges') === 'true') {
           history.push('/install');
         } else if (location.state && location.state.hasOwnProperty('from')) {
-            history.push(location.state.from.pathname);
+            history.push(location.state['from'].pathname);
         } else {
             history.push('/home');
         }
@@ -107,7 +107,7 @@ const App: React.FC<Props> = ({history, location}) => {
           <LoadData/>
         </PrivateRoute>
         <PrivateRoute path="/entity-tiles" exact>
-          <EntityTiles/>
+          <EntityTypes/>
         </PrivateRoute>
         <PrivateRoute path="/bench" exact>
           <Bench/>