+++ one-ui/ui/src/pages/EntityTypes.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import EntityTypes from './EntityTypes';
+
+describe('Install component', () => {
+  it('should render correctly', () => {
+    shallow(<EntityTypes />);
+  });
+});